requirements.txt
streamlit
python-dotenv
pypdf2
python-docx
streamlit-extras
langchain
openai
